# Conflicts:
#	CHANGELOG.md
#	data/ddl/table/STATUT.php
#	data/ddl/view/V_CONTRAT_MAIN.sql
#	data/ddl_columns_pos.php
#	module/Application/view/application/parametre/generaux.phtml
#	module/Intervenant/src/Entity/Db/Mapping/Intervenant.Entity.Db.Statut.dcm.xml
#	module/Intervenant/src/Entity/Db/Statut.php
#	module/Intervenant/src/Form/StatutSaisieForm.php

+++ module/Mission/src/Controller/SaisieController.php
@@ -145,8 +145,15 @@ class SaisieController extends AbstractController
 
         $form = $this->getFormMission();
 
+        if ($mission->isValide()){
+            $form->editValide();
+        }
+
         if ($this->getServiceContext()->getStructure()) {
             $form->remove('structure');
+            if (!$mission->getStructure()){
+                $mission->setStructure($this->getServiceContext()->getStructure());
+            }
         }
 
         $hDeb = $mission->getHeures();
@@ -186,15 +193,19 @@ class SaisieController extends AbstractController
         /** @var Mission $mission */
         $mission = $this->getEvent()->getParam('mission');
 
-        $this->getProcessusPlafond()->beginTransaction();
-        try {
-            $this->getServiceMission()->delete($mission);
-            $this->updateTableauxBord($mission);
-            $this->flashMessenger()->addSuccessMessage("Mission supprimée avec succès.");
-        } catch (\Exception $e) {
-            $this->flashMessenger()->addErrorMessage($this->translate($e));
+        if ($mission->canSupprimer()){
+            $this->getProcessusPlafond()->beginTransaction();
+            try {
+                $this->getServiceMission()->delete($mission);
+                $this->updateTableauxBord($mission);
+                $this->flashMessenger()->addSuccessMessage("Mission supprimée avec succès.");
+            } catch (\Exception $e) {
+                $this->flashMessenger()->addErrorMessage($this->translate($e));
+            }
+            $this->getProcessusPlafond()->endTransaction($mission, $typeVolumeHoraire, true);
+        }else{
+            $this->flashMessenger()->addErrorMessage('Vous n\'avez pas la possibilité de supprimer cette mission : elle a déjà été validée ou a fait l\'objet d\'un contrat');
         }
-        $this->getProcessusPlafond()->endTransaction($mission, $typeVolumeHoraire, true);
 
         return new AxiosModel([]);
     }
